Resolved Issue #47
On Hitting Sign In or Sign Up Button Without entering
any details make the app to crash.

I have done changes in MainActivity.java in line No. 103 and 153 by wrapping it into the condition so that when the app makes call to firebase authentication it always have email and password and if not pop up toast to ask for the same.